Move channel tests to own file
